almost ok multiple cast
