+++ ReleaseUpdate.md
@@ -1,0 +1,5 @@
+IPC between browser and golang to only manage one webrtc connection
+
+Webrtc from browser to browser (easier)
+
+No React / NodeJS (badd idea, too much for this small project)
