--- client_ui/public/wconfig.json
@@ -1,5 +1,0 @@
-{
-    "ip_address": "localhost",
-    "port": "3001",
-    "window_name": ""
-}
