--- webserver/admin/styles.css
@@ -1,25 +1,0 @@
-body{
-    margin: 0px;
-	padding: 0px;
-    /* width: 100%; */
-}
-
-.log-container {
-    margin: 4px;
-    padding: 4px;
-}
-
-table {
-    border: 1px solid #ccc;
-}
-
-th, td {
-    border: 1px solid #ccc;
-    font-size: 11px;
-}
-
-pre {
-    width: 50vw;
-    max-width: 600px;
-    overflow-x: auto;
-}
