--- webserver/client/defaultconfig.json
@@ -1,61 +1,0 @@
-[
-    {
-        "controller": "PS3",
-        "buttons": {
-            "0": [true, 5],
-            "1": [true, 6],
-            "2": [true, 7],
-            "3": [true, 4],
-            "4": [true, 0],
-            "5": [true, 3],
-            "6": [true, 1],
-            "7": [true, 2],
-            "8": [false, 3],
-            "9": [false, 4],
-            "10": [true, 8],
-            "11": [true, 9],
-            "12": [true, 15],
-            "13": [true, 13],
-            "14": [true, 12],
-            "15": [true, 14],
-            "16": [true, 10],
-            "17": [true, 17],
-            "18": [true, 18]
-        },
-        "analogButtons": {
-            "0": [false, 0],
-            "1": [false, 1],
-            "2": [false, 2],
-            "3": [false, 6],
-            "4": [false, 6],
-            "5": [false, 5]
-        }
-    },
-    {
-        "controller": "XBOX360",
-        "buttons": {
-            "0": [true, 12],
-            "1": [true, 13],
-            "2": [true, 14],
-            "3": [true, 15],
-            "4": [true, 8],
-            "5": [true, 9],
-            "6": [false, 3],
-            "7": [false, 4],
-            "8": [true, 10],
-            "9": [true, 4],
-            "10": [true, 6],
-            "12": [true, 0],
-            "13": [true, 1],
-            "14": [true, 2],
-            "15": [true, 3]
-        },
-        "analogButtons": {
-            "0": [false, 0],
-            "1": [false, 1],
-            "2": [false, 2],
-            "5": [false, 5]
-        }
-    }
-]
-
