--- webserver/client/wconfig.json
@@ -1,5 +1,0 @@
-{
-    "ip_address": "localhost",
-    "port": "3000",
-    "window_name": ""
-}
