+++ webserver/styles.css
@@ -1,0 +1,46 @@
+body{
+    margin: 0px;
+	padding: 0px;
+}
+
+video {
+    width: 100%;
+    height: 90%;
+    border: 2px solid #000;
+    object-fit: 'contain';
+}
+
+.tab {
+    display: flex;
+    justify-content: space-between;
+    width: 100%;
+}
+
+.tab button {
+    flex-grow: 1;
+}
+
+.chat-window {
+    width: 300px;
+    height: 400px;
+    border: 1px solid #ccc;
+    overflow-y: scroll;
+}
+
+.chat-message {
+    border-bottom: 1px solid #ccc;
+}
+
+.table-container {
+    display: flex;
+    border: 1px solid #ccc;
+    flex-wrap: wrap;
+}
+
+table {
+    border: 1px solid #ccc;
+}
+
+.selected-button {
+    fill: blue;
+}
