+++ webserver/webserver.go
@@ -2,30 +2,31 @@ package main
 
 import (
 	"net/http"
+
 	// "net/http/httputil"
 	// "net/url"
 	"log"
 	"os"
 
 	"bufio"
-	// "strings"
+
 	// "path/filepath"
 	"bytes"
 )
 
 var (
-	serverIP    string // IP address for WebSocket connection
-	targetPort  string // Port number for WebSocket connection
-	title       string // Title of the application to capture
+	serverIP   string // IP address for WebSocket connection
+	targetPort string // Port number for WebSocket connection
+	// title      string // Title of the application to capture
 )
 
-var logFilePaths = map[string]string{
-	"video":  "../logs/video.log",
-	"audio":   "../logs/audio.log",
-	"client":  "../logs/client.log",
-	"server":   "../logs/server.log",
-	"sign":  "../logs/sign.log",
-}
+// var logFilePaths = map[string]string{
+// 	"video":  "../logs/video.log",
+// 	"audio":  "../logs/audio.log",
+// 	"client": "../logs/client.log",
+// 	"server": "../logs/server.log",
+// 	"sign":   "../logs/sign.log",
+// }
 
 /**
  * main is the main entry point of the Webserver program. It sets up HTTP handlers to serve static files,
@@ -44,24 +45,24 @@ func main() {
 	log.Println(serverIP)
 	log.Println(targetPort)
 
-	serverPort := "80"
+	// serverPort := "80"
 	// streamTarget := "http://" + serverIP + ":" + targetPort
-	
+
 	// Serve the static files from the React build directory
-	http.Handle("/", http.FileServer(http.Dir("../ui/build")))
+	// http.Handle("/", http.FileServer(http.Dir("../ui/build")))
 
 	// Start the server
-	addr := serverIP + ":" + serverPort
+	addr := serverIP + ":" + targetPort
 	log.Printf("Server listening on %s\n", addr)
 	// log.Fatal(http.ListenAndServe(addr, nil))
 
-	// // Handle requests for static files for the client page
-	// http.Handle("/", http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-	// 	log.Printf("Client request: %s %s\n", r.Method, r.URL.Path)
-	// 	http.FileServer(http.Dir("client")).ServeHTTP(w, r)
-	// }))
+	// Handle requests for static files for the client page
+	http.Handle("/", http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		log.Printf("Client request: %s %s\n", r.Method, r.URL.Path)
+		http.FileServer(http.Dir("./")).ServeHTTP(w, r)
+	}))
 
-	// Handle requests for static files for the admin page
+	// // Handle requests for static files for the admin page
 	// http.Handle("/admin/", http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 	// 	log.Printf("Admin request: %s %s\n", r.Method, r.URL.Path)
 	// 	http.StripPrefix("/admin/", http.FileServer(http.Dir("admin"))).ServeHTTP(w, r)
@@ -75,61 +76,60 @@ func main() {
 	// 	proxy.ServeHTTP(w, r)
 	// })
 
-	// Add a new handler for serving the log file content to the admin page
-	http.HandleFunc("/admin/logs/", func(w http.ResponseWriter, r *http.Request) {
-		// Extract the log type from the request URL
-		logType := strings.TrimPrefix(r.URL.Path, "/admin/logs/")
-	
-		log.Printf("Admin log request: %s %s (Log type: %s)\n", r.Method, r.URL.Path, logType)
-	
-		// Get the log file path based on the log type
-		logFilePath, ok := logFilePaths[logType]
-		if !ok {
-			http.Error(w, "Invalid log type", http.StatusBadRequest)
-			return
-		}
+	// // Add a new handler for serving the log file content to the admin page
+	// http.HandleFunc("/admin/logs/", func(w http.ResponseWriter, r *http.Request) {
+	// 	// Extract the log type from the request URL
+	// 	logType := strings.TrimPrefix(r.URL.Path, "/admin/logs/")
 
-		log.Println(logFilePath)
+	// 	log.Printf("Admin log request: %s %s (Log type: %s)\n", r.Method, r.URL.Path, logType)
 
-		absPath, err := filepath.Abs(logFilePath)
-		if err != nil {
-			http.Error(w, "Can't find file path", http.StatusBadRequest)
-			return
-		}
+	// 	// Get the log file path based on the log type
+	// 	logFilePath, ok := logFilePaths[logType]
+	// 	if !ok {
+	// 		http.Error(w, "Invalid log type", http.StatusBadRequest)
+	// 		return
+	// 	}
 
-		// Open the log file
-		file, err := os.Open(absPath)
-		if err != nil {
-			http.Error(w, "Error opening log file", http.StatusInternalServerError)
-			return
-		}
-		defer file.Close()
+	// 	log.Println(logFilePath)
 
-		// Read the last 10 lines of the log file
-		lines, err := readLastLines(file, 10)
-		if err != nil {
-			http.Error(w, "Error reading log file", http.StatusInternalServerError)
-			return
-		}
+	// 	absPath, err := filepath.Abs(logFilePath)
+	// 	if err != nil {
+	// 		http.Error(w, "Can't find file path", http.StatusBadRequest)
+	// 		return
+	// 	}
 
-		// Set the content type as "text/plain" to display plain text content
-		w.Header().Set("Content-Type", "text/plain")
+	// 	// Open the log file
+	// 	file, err := os.Open(absPath)
+	// 	if err != nil {
+	// 		http.Error(w, "Error opening log file", http.StatusInternalServerError)
+	// 		return
+	// 	}
+	// 	defer file.Close()
 
-		// Write the log file content to the response
-		for _, line := range lines {
-			_, err := w.Write([]byte(line + "\n"))
-			if err != nil {
-				http.Error(w, "Error writing response", http.StatusInternalServerError)
-				return
-			}
-		}
-	})
+	// 	// Read the last 10 lines of the log file
+	// 	lines, err := readLastLines(file, 10)
+	// 	if err != nil {
+	// 		http.Error(w, "Error reading log file", http.StatusInternalServerError)
+	// 		return
+	// 	}
+
+	// 	// Set the content type as "text/plain" to display plain text content
+	// 	w.Header().Set("Content-Type", "text/plain")
+
+	// 	// Write the log file content to the response
+	// 	for _, line := range lines {
+	// 		_, err := w.Write([]byte(line + "\n"))
+	// 		if err != nil {
+	// 			http.Error(w, "Error writing response", http.StatusInternalServerError)
+	// 			return
+	// 		}
+	// 	}
+	// })
 
 	// http.ListenAndServe(serverIP+":"+serverPort, nil)
 
 	log.Fatal(http.ListenAndServe(addr, nil))
 }
-
 
 /**
  * readLastLines reads the last 'n' lines from an open file, handling various line endings.
@@ -174,4 +174,3 @@ func readLastLines(file *os.File, n int) ([]string, error) {
 
 	return lines[startIdx:], nil
 }
-
